job #5007 reorganized the three test suites
