+++ src/org.xtuml.bp.debug.ui.test/generate.xml
@@ -10,7 +10,7 @@
   <target name="nb_all" description="Generate all code for this project" depends="generate_set_operations_tests" />
 
   <property name="set_op_simple" value="matrices/set_operations_simple_matrix.txt" />
-  <property name="set_op_simple_output" value="src/org/xtuml/bp/debug/ui/test/sets/SetOperationsTests_0.java" />
+  <property name="set_op_simple_output" value="src/org/xtuml/bp/debug/ui/test/sets/SetOperationsSimpleTests_0.java" />
   <property name="set_op_precedence" value="matrices/set_operations_precedence_matrix.txt" />
   <property name="set_op_precedence_output" value="src/org/xtuml/bp/debug/ui/test/sets/SetOperationsPrecedenceTests_0.java" />
   <property name="set_op_compare" value="matrices/set_operations_compare_matrix.txt" />
@@ -23,7 +23,7 @@
 
   <target name="generate_set_operations_tests" description="Generate test code set operations" depends="set_operations.depends" unless="set_operations.buildNotNeeded">
     <exec vmlauncher="false" failonerror="true" executable="perl">
-      <arg line="../org.xtuml.bp.test/UnitTestGenerator.pl ${set_op_simple} src/org/xtuml/bp/debug/ui/test/sets/SetOperationsTests.java
+      <arg line="../org.xtuml.bp.test/UnitTestGenerator.pl ${set_op_simple} src/org/xtuml/bp/debug/ui/test/sets/SetOperationsSimpleTests.java
         -n 250 -p org.xtuml.bp.debug.ui.test.sets -suite -DNO -data -noResults" />
     </exec>
     <exec vmlauncher="false" failonerror="true" executable="perl">
@@ -44,6 +44,7 @@
     <delete>
       <fileset dir="src/org/xtuml/bp/debug/ui/test/sets/" >
         <exclude name="SetOperationsTests.java" />
+        <exclude name="SetOperationsSimpleTests.java" />
         <exclude name="SetOperationsPrecedenceTests.java" />
         <exclude name="SetOperationsCompareTests.java" />
         <include name="**.java" />
