+++ src/org.xtuml.bp.debug.ui.test/src/org/xtuml/bp/debug/ui/test/sets/SetOperationsSimpleTests.java
@@ -2,66 +2,66 @@ package org.xtuml.bp.debug.ui.test.sets;
 
 import org.xtuml.bp.core.common.NonRootModelElement;
 
-public class SetOperationsCompareTests extends SetOperationsTests {
+public class SetOperationsSimpleTests extends SetOperationsTests {
 
     public static boolean generateResults = false;
     public static boolean useDrawResults = true;
 
-    public SetOperationsCompareTests( String subTypeClassName, String subTypeArg0 ) {
+    public SetOperationsSimpleTests( String subTypeClassName, String subTypeArg0 ) {
         super( subTypeClassName, subTypeArg0 );
     }
 
     /**
-     * "H" is one of the degrees of freedom as specified in this issues
+     * "O" is one of the degrees of freedom as specified in this issues
      * test matrix.
-     * This routine gets the "H" instance from the given name.
+     * This routine gets the "O" instance from the given name.
      * 
      * @param element The degree of freedom instance to retrieve
      * @return A model element used in the test as specified by the test matrix
      */
-    NonRootModelElement selectH( String element ) {
-        return selectH( element, null );
+    public NonRootModelElement selectO( String element ) {
+        return selectO( element, null );
     }
 
     /**
-     * "H" is one of the degrees of freedom as specified in this issues
+     * "O" is one of the degrees of freedom as specified in this issues
      * test matrix.
-     * This routine gets the "H" instance from the given name.
+     * This routine gets the "O" instance from the given name.
      * 
      * @param element The degree of freedom instance to retrieve
      * @param extraData Extra data needed for selection
      * @return A model element used in the test as specified by the test matrix
      */
-    NonRootModelElement selectH( String element, Object extraData ) {
+    public NonRootModelElement selectO( String element, Object extraData ) {
         NonRootModelElement nrme = selectColumn( element, extraData );
-        assertTrue("An instance with degree of freedom type \"H\" was not found.  Instance Name: " + element + ".", nrme!=null);
+        assertTrue("An instance with degree of freedom type \"O\" was not found.  Instance Name: " + element + ".", nrme!=null);
         return nrme;
     }
 
     /**
-     * "IJ" is one of the degrees of freedom as specified in this issues
+     * "AB" is one of the degrees of freedom as specified in this issues
      * test matrix.
-     * This routine gets the "IJ" instance from the given name.
+     * This routine gets the "AB" instance from the given name.
      * 
      * @param element The degree of freedom instance to retrieve
      * @return A model element used in the test as specified by the test matrix
      */
-    NonRootModelElement selectIJ( String element ) {
-        return selectIJ( element, null );
+    public NonRootModelElement selectAB( String element ) {
+        return selectAB( element, null );
     }
 
     /**
-     * "IJ" is one of the degrees of freedom as specified in this issues
+     * "AB" is one of the degrees of freedom as specified in this issues
      * test matrix.
-     * This routine gets the "IJ" instance from the given name.
+     * This routine gets the "AB" instance from the given name.
      * 
      * @param element The degree of freedom instance to retrieve
      * @param extraData Extra data needed for selection
      * @return A model element used in the test as specified by the test matrix
      */
-    NonRootModelElement selectIJ( String element, Object extraData ) {
+    public NonRootModelElement selectAB( String element, Object extraData ) {
         NonRootModelElement nrme = selectRow( element, extraData );
-        assertTrue("An instance with degree of freedom type \"IJ\" was not found.  Instance Name: " + element + ".", nrme!=null);
+        assertTrue("An instance with degree of freedom type \"AB\" was not found.  Instance Name: " + element + ".", nrme!=null);
         return nrme;
     }
 
@@ -73,7 +73,7 @@ public class SetOperationsCompareTests extends SetOperationsTests {
      * @param columnInstance Model instance from the column
      * @param rowInstance Model instance from the row
      */
-    void H_IJ_Action( NonRootModelElement columnInstance, NonRootModelElement rowInstance ) {
+    public void O_AB_Action( NonRootModelElement columnInstance, NonRootModelElement rowInstance ) {
         testAction( columnInstance, rowInstance );
     }
